Add files via upload
Impulse Voting Project is created in both application & web-based
a system that will help you manage your elections easily and securely.
This voting system can be used for casting votes during the elections
held in colleges, events, etc. A database is being maintained in which
all the name of the voters with their complete information is stored
along with the Face Recognition Mechanism.